# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Giang API",
        "description": "description",
        "contact": {
            "email": "giangact.dev@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://giang.stayhere.tk/api",
            "description": "Staging API v1"
        },
        {
            "url": "http://giang.test:8080/api",
            "description": "Local test API v1"
        }
    ],
    "paths": {
...
        "/register/otp": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "",
                "description": "",
                "operationId": "sendOtpForRegister",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/driver/login": {
            "post": {
                "tags": [
                    "driver"
                ],
                "summary": "",
                "description": "",
                "operationId": "loginWithOtp",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/driver/login/otp": {
            "post": {
                "tags": [
                    "driver"
                ],
                "summary": "",
                "description": "",
                "operationId": "loginsendOtp",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/driver/register": {
            "post": {
                "tags": [
                    "driver"
                ],
                "summary": "",
                "description": "",
                "operationId": "registerdriver",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/driver/register/otp": {
            "post": {
                "tags": [
                    "driver"
                ],
                "summary": "",
                "description": "",
                "operationId": "sendOtpForRegister",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/driver/orders/summary": {
            "get": {
                "tags": [
                    "driver"
                ],
                "summary": "",
                "description": "",
                "operationId": "summary",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}